requirements.txt
Django==4.2
djangorestframework==3.14.0
phonenumbers==8.13.8
djangorestframework-simplejwt==5.2.2